Rename partition command for VM partitions
